requirements/main.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements/main.in
#
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
chardet==4.0.0
    # via pdfminer-six
charset-normalizer==2.0.12
    # via requests
commonmark==0.9.1
    # via rich
cryptography==36.0.2
    # via pdfminer-six
idna==3.3
    # via requests
numpy==1.22.3
    # via -r requirements/main.in
pdfminer-six==20220319
    # via
    #   -r requirements/main.in
    #   pdfplumber
pdfplumber==0.6.2
    # via -r requirements/main.in
pillow==9.1.0
    # via pdfplumber
pycparser==2.21
    # via cffi
pygments==2.12.0
    # via rich
pypdf2==1.27.12
    # via -r requirements/main.in
requests==2.27.1
    # via
    #   -r requirements/main.in
    #   tika
rich==12.4.0
    # via -r requirements/main.in
tika==1.24
    # via -r requirements/main.in
urllib3==1.26.9
    # via requests
wand==0.6.7
    # via pdfplumber

# The following packages are considered to be unsafe in a requirements file:
# setuptools
